Criação Enum para estrelas dos personagens Criação de retorno de personagem random Memory cache

+++ Domain/DTO/Common/AccountDto.cs
@@ -4,6 +4,6 @@ namespace Domain.DTO.Common;
 
 public class AccountDto
 {
-    public int AccountId { get; set; }
+    public ulong AccountId { get; set; }
     public List<GenshinCharacterDto> GenshinCharacters { get; set; }
 }
